api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
moviepy==1.0.3
textblob==0.19.0
deep-translator==1.11.4
groq==0.30.0
nltk==3.9.1
python-dotenv==1.0.0

